Split out activateTheme
